+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   Add `<twig:Turbo:Stream>` component
 -   Add `<twig:Turbo:Frame>` component
+-   Add support for custom actions in `TurboStream` and `TurboStreamResponse`
 
 ## 2.21.0